+++ test_package/find_package/main.cpp
@@ -1,5 +1,9 @@
 #include "FbSingleAxis.h"
+#include "Global.h"
 #include "Scheduler.h"
+
+static_assert(plcopen::MC_SERVO_EXTENSION_DIGITAL_INPUT_BASE == 10000);
+static_assert(plcopen::MC_SERVO_EXTENSION_DIGITAL_OUTPUT_BASE == 20000);
 
 int main()
 {